--- a/go.mod
+++ b/go.mod
@@ -11,5 +11,3 @@ require (
 	golang.org/x/net v0.0.0-20190206173232-65e2d4e15006 // indirect
 	golang.org/x/sys v0.0.0-20190204203706-41f3e6584952 // indirect
 )
-
-go 1.12