--- a/prometheus/label.go
+++ b/prometheus/label.go
@@ -1,6 +1,8 @@
 package prometheus
 
 import (
+	"bytes"
+
 	"github.com/segmentio/fasthash/jody"
 	"github.com/segmentio/stats"
 )
@@ -17,12 +19,12 @@ func (l label) hash() uint64 {
 	return h
 }
 
-func (l1 label) equal(l2 label) bool {
-	return l1.name == l2.name && l1.value == l2.value
+func (l label) equal(l2 label) bool {
+	return l.name == l2.name && l.value == l2.value
 }
 
-func (l1 label) less(l2 label) bool {
-	return l1.name < l2.name || (l1.name == l2.name && l1.value < l2.value)
+func (l label) less(l2 label) bool {
+	return l.name < l2.name || (l.name == l2.name && l.value < l2.value)
 }
 
 type labels []label
@@ -31,6 +33,19 @@ func makeLabels(l ...label) labels {
 	m := make(labels, len(l))
 	copy(m, l)
 	return m
+}
+
+func (l labels) filterNamed(names []byte) labels {
+	if len(names) == 0 {
+		return l
+	}
+	out := make(labels, 0, len(l))
+	for i := range l {
+		if !bytes.Contains(names, []byte(l[i].name)) {
+			out = append(out, l[i])
+		}
+	}
+	return out
 }
 
 func (l labels) copyAppend(m ...label) labels {
@@ -55,25 +70,25 @@ func (l labels) hash() uint64 {
 	return h
 }
 
-func (l1 labels) equal(l2 labels) bool {
-	if len(l1) != len(l2) {
+func (l labels) equal(l2 labels) bool {
+	if len(l) != len(l2) {
 		return false
 	}
-	for i := range l1 {
-		if !l1[i].equal(l2[i]) {
+	for i := range l {
+		if !l[i].equal(l2[i]) {
 			return false
 		}
 	}
 	return true
 }
 
-func (l1 labels) less(l2 labels) bool {
-	n1 := len(l1)
+func (l labels) less(l2 labels) bool {
+	n1 := len(l)
 	n2 := len(l2)
 
 	for i := 0; i != n1 && i != n2; i++ {
-		if !l1[i].equal(l2[i]) {
-			return l1[i].less(l2[i])
+		if !l[i].equal(l2[i]) {
+			return l[i].less(l2[i])
 		}
 	}