--- a/prometheus/label_test.go
+++ b/prometheus/label_test.go
@@ -1,6 +1,9 @@
 package prometheus
 
-import "testing"
+import (
+	"reflect"
+	"testing"
+)
 
 func TestLabelsLess(t *testing.T) {
 	tests := []struct {
@@ -55,6 +58,59 @@ func TestLabelsLess(t *testing.T) {
 		t.Run("", func(t *testing.T) {
 			if less := test.l1.less(test.l2); less != test.less {
 				t.Errorf("(%#v < %#v) != %t", test.l1, test.l2, test.less)
+			}
+		})
+	}
+}
+
+func TestFilter(t *testing.T) {
+	var testFilter = []byte{'l', '1', 0x00, 'l', '2'}
+	var testLabels = labels{
+		{
+			name: "l1",
+		},
+		{
+			name: "l2",
+		},
+		{
+			name: "l3",
+		},
+		{
+			name: "l4",
+		},
+	}
+	tests := []struct {
+		name   string
+		in     labels
+		filter []byte
+		expect labels
+	}{
+		{
+			name:   "no labels",
+			in:     labels(nil),
+			filter: testFilter,
+			expect: make(labels, 0),
+		},
+		{
+			name:   "no filters",
+			in:     testLabels,
+			filter: []byte(nil),
+			expect: testLabels,
+		},
+		{
+			name: "active filters",
+			in: testLabels,
+			filter:testFilter,
+			expect: labels{{name: "l3"}, {name: "l4"}},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			l := test.in.filterNamed(test.filter)
+
+			if !reflect.DeepEqual(l, test.expect) {
+				t.Errorf("\nexpected: %#v\n     got: %#v", test.expect, l)
 			}
 		})
 	}