This allows the prometheus handler to filter out high-cardinality labels (e.g. `http_req_path`) from observations

Usage:
```go
// create prometheus handler
handler := &prometheus.Handler{
    TrimPrefix: cfg.StatsDPrefix,
}
// set HC tags
handler.IgnoreLabels([]string{"http_req_path"})
```

Benchmarks:

Before filter (current master):

```
goos: darwin
goarch: amd64
pkg: github.com/segmentio/stats/prometheus
BenchmarkHandleMetric/counter-8                  5687409               189 ns/op               0 B/op          0 allocs/op
BenchmarkHandleMetric/gauge-8                    6038472               187 ns/op               0 B/op          0 allocs/op
BenchmarkHandleMetric/histogram-8                5275975               230 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/segmentio/stats/prometheus   4.078s

```

After filter (this branch):

```
goos: darwin
goarch: amd64
pkg: github.com/segmentio/stats/prometheus
BenchmarkHandleMetric/counter-8                  6561445               186 ns/op               0 B/op          0 allocs/op
BenchmarkHandleMetric/gauge-8                    6449792               184 ns/op               0 B/op          0 allocs/op
BenchmarkHandleMetric/histogram-8                5353644               223 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/segmentio/stats/prometheus   4.224s

```