@schigh 2019-09-17
I grabbed a random assortment of gophers to look at this PR.
Once we approve, I will make an upstream PR.